# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Proyectos y Canales",
    "version": "1.0.0",
    "description": "API para gestionar proyectos, canales y autenticación"
  },
  "paths": {
...
    "/api/comment": {
      "get": {
        "summary": "Obtiene una lista de comentarios de una tarea",
        "tags": ["Comment"],
        "parameters": [
          {
            "in": "query",
            "name": "taskId",
            "schema": { "type": "string" },
            "required": true,
            "description": "ID de la tarea a la que pertenecen los comentarios"
          },
          {
            "in": "query",
            "name": "search",
            "schema": { "type": "string" },
            "description": "Término de búsqueda opcional para filtrar comentarios"
          },
          {
            "in": "query",
            "name": "page",
            "schema": { "type": "integer" },
            "description": "Número de página para paginación"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": { "type": "integer" },
            "description": "Cantidad de comentarios por página"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de comentarios obtenida correctamente"
          },
          "400": { "description": "Falta el taskId" },
          "401": { "description": "No autorizado" },
          "403": { "description": "Acceso no autorizado" },
          "404": { "description": "Tarea no encontrada" },
          "500": { "description": "Error al recuperar comentarios" }
        }
      },
      "post": {
        "summary": "Crea un nuevo comentario en una tarea",
        "tags": ["Comment"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "taskId": {
                    "type": "string",
                    "description": "ID de la tarea donde se agregará el comentario"
                  },
                  "content": {
                    "type": "string",
                    "description": "Contenido del comentario"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Comentario creado correctamente" },
          "401": { "description": "No autorizado" },
          "403": {
            "description": "No tienes permisos suficientes para comentar"
          },
          "404": { "description": "Tarea no encontrada" },
          "500": { "description": "Error al crear el comentario" }
        }
      },
      "put": {
        "summary": "Actualiza un comentario existente",
        "tags": ["Comment"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "_id": {
                    "type": "string",
                    "description": "ID del comentario a actualizar"
                  },
                  "content": {
                    "type": "string",
                    "description": "Nuevo contenido del comentario"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Comentario actualizado correctamente" },
          "401": { "description": "No autorizado" },
          "403": {
            "description": "Solo el autor puede editar este comentario"
          },
          "404": { "description": "Comentario no encontrado" },
          "500": { "description": "Error al actualizar el comentario" }
        }
      },
      "delete": {
        "summary": "Elimina un comentario",
        "tags": ["Comment"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "_id": {
                    "type": "string",
                    "description": "ID del comentario a eliminar"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Comentario eliminado exitosamente" },
          "401": { "description": "No autorizado" },
          "403": {
            "description": "No tienes permiso para eliminar este comentario"
          },
          "404": { "description": "Comentario no encontrado" },
          "500": { "description": "Error al eliminar el comentario" }
        }
      }
    },
...
    "/api/projects": {
      "get": {
        "summary": "Obtiene una lista de proyectos",
        "tags": ["Project"],
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "schema": { "type": "string" },
            "description": "Texto de búsqueda para filtrar proyectos por nombre o descripción"
          },
          {
            "in": "query",
            "name": "sort",
            "schema": { "type": "string" },
            "description": "Campo por el cual ordenar los resultados"
          },
          {
            "in": "query",
            "name": "page",
            "schema": { "type": "integer" },
            "description": "Número de página para la paginación"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": { "type": "integer" },
            "description": "Cantidad de proyectos por página"
          }
        ],
        "responses": {
          "200": { "description": "Lista de proyectos obtenida exitosamente" },
          "401": { "description": "No autorizado" },
          "500": { "description": "Error del servidor" }
        }
      },
      "post": {
        "summary": "Crea un nuevo proyecto",
        "tags": ["Project"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "projectName": { "type": "string" },
                  "background": { "type": "string" },
                  "visibility": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Proyecto creado exitosamente" },
          "401": { "description": "No autorizado" },
          "500": { "description": "Error del servidor" }
        }
      },
      "put": {
        "summary": "Actualiza un proyecto existente",
        "tags": ["Project"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "_id": { "type": "string" },
                  "name": { "type": "string" },
                  "backgroundType": { "type": "string" },
                  "backgroundColor": { "type": "string" },
                  "backgroundGradient": {
                    "type": "object",
                    "properties": {
                      "color1": { "type": "string" },
                      "color2": { "type": "string" },
                      "angle": { "type": "number" }
                    }
                  },
                  "backgroundImage": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Proyecto actualizado exitosamente" },
          "401": { "description": "No autorizado" },
          "403": {
            "description": "No tienes permiso para editar este proyecto"
          },
          "500": { "description": "Error del servidor" }
        }
      },
      "delete": {
        "summary": "Elimina un proyecto",
        "tags": ["Project"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "_id": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Proyecto eliminado exitosamente" },
          "401": { "description": "No autorizado" },
          "403": {
            "description": "No tienes permiso para eliminar este proyecto"
          },
          "500": { "description": "Error del servidor" }
        }
      }
    },
...
  }
}